add auth

+++ controllers/cards.js
@@ -16,10 +16,14 @@ module.exports.createCard = (req, res) => {
 };
 
 module.exports.deleteCard = (req, res) => {
-  Card.findByIdAndRemove(req.params.cardId)
-    .orFail()
-    .then((card) => res.send({ data: card }))
-    .catch((err) => errorMessage(err, req, res));
+  if (req.params.cardId === req.user._id) {
+    Card.findByIdAndRemove(req.params.cardId)
+      .orFail()
+      .then((card) => res.send({ data: card }))
+      .catch((err) => errorMessage(err, req, res));
+  } else {
+    res.status(401).send({ message: 'Нельзя удалять чужие карточки' });
+  }
 };
 
 module.exports.likeCard = (req, res) => {
